Fix validation of operands in unreachable context
Even though, logically, it is a noop to pop an element from the operand
stack and push it back right after, it is necessary to do so to
correctly validate code after an inconditional break occurred.
In such a case, popping an element is always possible and the stack may
be viewed as empty whereas it is not.

For instance, let’s assume we are in an unreachable contextand the
operand stack consists in the single element `i32`.
If we try to validate an `i32.add` instruction in such context, the
validation succeeds because when unreachable, we can pop whatever we
want from an empty stack. However, we must explicitly mark the top
element as `i32` after validating this instruction. This is why we
always pop and push back when validating, to ensure, corner cases are
correctly handled.