Fix validation when code section is missing
If function types where defined and code section is missing, then the
module is not valid as the number of defined signatures and codes are
different.
To be sure to handle that case correctly, we must somehow know when the
stream is over to perform that last check. Using the fs2 couple
`nonTerminate`/`unNoneTerminate` we can perform this last check before
stream ends.