Wrap constructors in effectful type
Any Swam element constructor might depend on some configuration or
resource allocation. Even if they currently don't, they might need it in
the future. To prepare for that, make these constructors private, and
provide an `apply` method in the companion object that returns the new
instance in an effectful type. This was already the case for the engine,
make it for other compilers and decompilers part of the public API.