Implement low level stack more efficiently
The high level interpreter creates a lot of objects whenever a new frame
is pushed, or even at every interpretation step. Whereas this is ok for
simple scripts, it is quite costly and adds a lot of overhead at
runtime. The idea is to make pushing frames object creation free in the
end. This lower level representation of a thread frame, is one step
towards that result.