Make low level interpreter unsafe
Being unsafe, means calling side-effecting functions, not wrapped inside
effectful types. This is necessary to be gentle on memory allocations
that otherwise happen at any interpretation step. While this is not a
problem with short simple computations, it becomes quite heavy when
dealing with memories.

To mitigate that, having unsafe internals that execute unsafe code but
are completely controlled by the interpreter, offers a still safe
interface with better performance characteristics.