Fix bug in opcode decoding
If the input has less than a byte available, then it must fail to decode
the opcode. The `splitAt` method won’t fail if the index is greater than
the BitVector length, and just return the input as prefix and the empty
vector as suffix. This results in reading an infinite amount of
`Unreachable` instructions if the input end prematurely when reading a
code block.